chore: reorganize classes
