+++ lib/classes/command.dart
@@ -1,0 +1,4 @@
+class Command {
+  final Map<String, List<String>> commands;
+  const Command({required this.commands});
+}
