+++ lib/classes/command_result.dart
@@ -1,0 +1,13 @@
+class CommandResult {
+  final String command;
+  final String stdout;
+  final String stderr;
+  final int exitCode;
+
+  const CommandResult({
+    required this.command,
+    required this.stdout,
+    this.stderr = "",
+    this.exitCode = 0,
+  });
+}
