+++ lib/classes/stage.dart
@@ -1,0 +1,6 @@
+import 'command.dart';
+
+class Stage {
+  final Map<String, Command> stages;
+  const Stage({required this.stages});
+}
