+++ lib/extensions/colored_string.dart
@@ -5,8 +5,8 @@ extension ColoredStringExtension on String {
   static const _yellow = "\x1B[33m";
   static const _reset = "\x1B[0m";
 
-  String get red => '$_red$this$_reset';
-  String get green => '$_green$this$_reset';
   String get blue => '$_blue$this$_reset';
+  String get green => '$_green$this$_reset';
+  String get red => '$_red$this$_reset';
   String get yellow => '$_yellow$this$_reset';
 }
